Tests refactor to be synchronized with the new factories hierarchy
